[wails] rename service -
