+++ internal/app/grpc.go
@@ -1,4 +1,4 @@
-package service
+package app
 
 import (
 	"bytes"
