+++ internal/app/handlers.go
@@ -1,4 +1,4 @@
-package service
+package app
 
 import (
 	"encoding/json"
