+++ internal/app/send_jq.go
@@ -1,4 +1,4 @@
-package service
+package app
 
 import (
 	"context"
