+++ internal/app/send_sql.go
@@ -1,4 +1,4 @@
-package service
+package app
 
 import (
 	"database/sql"
